+++ .gitignore
@@ -1,0 +1,5 @@
+*.pyc
+.env
+anime/
+test.py
+.vscode/
