+++ bots/anime_birthday_bot.py
@@ -1,7 +1,8 @@
 import tweepy
 import logging
-from config import create_api
 import time
+
+from configs.conf import create_api
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger()
@@ -10,7 +11,7 @@ logger = logging.getLogger()
 def tweet_birthday(api):
     logger.info('Tweeting birthdays')
     # api.update_status('This is my initial tweet. Hello world!')
-    api.update_status('This is tweet is from a bot. This is a test tweet!')
+    api.update_status('New tweet!')
 
 
 def main():
