--- bots/config.txt
@@ -1,4 +1,0 @@
-export CONSUMER_KEY="ZWDrd0Aa80p1oEd6dUAs96sNk"
-export CONSUMER_SECRET="kqZDbG8fnSGX1G1khgzl1tsiW8RlH9NtafHrJ5vd8vL6w2hLl4"
-export ACCESS_TOKEN="1211469743493746688-2ppf6rBRwLSGe8UQBpzLxc8ReIPnUE"
-export ACCESS_TOKEN_SECRET="UlitYRU0VY04ecAhVopcksUyQNOYDgt9DqErBEMphzO8t"
