+++ requirements.txt
@@ -7,4 +7,5 @@ requests==2.22.0
 requests-oauthlib==1.3.0
 six==1.13.0
 tweepy==3.8.0
-urllib3==1.25.7
+urllib3==1.25.8
+python-decouple==3.4
